Move `JavaRenderer` under the `render` package
