+++ tests/protodata-extension/src/main/java/io/spine/protodata/test/annotation/AnnotationRenderer.java
@@ -26,7 +26,7 @@
 
 package io.spine.protodata.test.annotation;
 
-import io.spine.protodata.java.JavaRenderer;
+import io.spine.protodata.java.render.JavaRenderer;
 import io.spine.protodata.render.SourceFileSet;
 import io.spine.protodata.test.Annotated;
 import io.spine.protodata.test.FieldId;
